Ensure 'default' is in the list of mate-terminal profiles
"dconf read" returns an empty string if no custom value has been set for
a particular key, even if a default value is specified for that key's
schema.

In the mate-terminal profiles list, the default profile value is
['default'] which would be overwritten when initially setting
mate-terminal configuration values using the mate-terminal task. To work
around this, 'default' is now explicitly added to the list of
mate-terminal profiles if it is not present.