Override gnome-terminal launcher for "Open in Terminal" functionality
gnome-terminal was previously configured to not preserve its working
directory on exit. This behavior is still desired, but is incompatible
with Nemo's "Open in Terminal" action.

This change overrides the gnome-terminal settings to preserve the
working directory which allows Nemo's "Open in Terminal" to work
properly. Additionally, the launcher is customized to specify a working
directory of `$HOME` by default, which ensures manual terminal startup
begins in the home directory.